backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]
websockets
pydantic>=2.0
anyio>=4.0.0
cryptography>=43.0.1
sqlalchemy
alembic
httpx
psycopg2-binary
pytest
cryptography
langgraph
langchain
langchain-core>=1.0.0
langchain-groq
python-dotenv
typing-extensions
pgvector
sentence_transformers
langchain-community
langchain-huggingface
faiss-cpu
unstructured
python-pptx
python-docx
pdfminer.six
openpyxl
langchain-experimental
langchain-cerebras
pypdf
docx2txt
wikipedia
browser-use>=0.9.5
langchain-openai>=1.0.0
playwright>=1.40.0
yfinance
pandas
Pillow
groq
requests
aiofiles
apscheduler>=3.10.0
mcp>=1.0.0
httpx-sse>=0.4.0
composio>=0.5.0
